Move decode_merge_name to business.py
